improved apimetrics regex and added test cases
